arraylist tidy up and refactor, switched to more expressive types

+++ include/arraylist.h
@@ -1,30 +1,33 @@
 #ifndef _ARRAYLIST_H
 #define _ARRAYLIST_H
 
+#include <stdint.h>
+#include <stdlib.h>
+
 struct arraylist {
-	unsigned _array_size;
-	unsigned _pos;
-	unsigned size;
+	size_t _array_size;
+	unsigned long _pos;
+	size_t size;
 	void **list;
 };
 
-int arraylist_init(struct arraylist*);
+uint8_t arraylist_init(struct arraylist*);
 void arraylist_destroy(struct arraylist*);
 void arraylist_destroy_free(struct arraylist*);
-int arraylist_add(struct arraylist*,void*);
-int arraylist_set(struct arraylist*,unsigned,void*);
-void *arraylist_get(struct arraylist*,unsigned);
+uint8_t arraylist_add(struct arraylist*,void*);
+uint8_t arraylist_set(struct arraylist*,unsigned long,void*);
+void *arraylist_get(struct arraylist*,unsigned long);
 void *arraylist_find(struct arraylist*,const void*,
 		int(*)(const void*,const void*));
-int arraylist_size(struct arraylist*);
-int arraylist_maxidx(struct arraylist*);
-int arraylist_indexof(struct arraylist*,const void*,
+size_t arraylist_size(struct arraylist*);
+size_t arraylist_maxidx(struct arraylist*);
+long  arraylist_indexof(struct arraylist*,const void*,
 		int(*)(const void*,const void*));
-int arraylist_contains(struct arraylist*,const void*,
+uint8_t arraylist_contains(struct arraylist*,const void*,
 		int(*)(const void*,const void*));
-int arraylist_remove(struct arraylist*,const void*,
+uint8_t arraylist_remove(struct arraylist*,const void*,
 		int(*)(const void*,const void*));
 void *arraylist_removeat(struct arraylist*,unsigned);
-int arraylist_copy(struct arraylist*,struct arraylist*);
+uint8_t arraylist_copy(struct arraylist*,struct arraylist*);
 
 #endif
